requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
httpx==0.27.2
cryptography==43.0.3
celery==5.4.0
redis==5.2.0
python-dateutil==2.9.0.post0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
